Update readme, changelog, examples
